Rename the module that now contains only GenKeys
