+++ src/lib.rs
@@ -38,6 +38,7 @@ pub mod entry;
 pub mod erasure;
 pub mod fetch_stage;
 pub mod fullnode;
+pub mod gen_keys;
 pub mod genesis_block;
 pub mod gossip_service;
 pub mod leader_scheduler;
@@ -57,7 +58,6 @@ pub mod rpc_pubsub;
 pub mod rpc_request;
 pub mod runtime;
 pub mod service;
-pub mod signature;
 pub mod sigverify;
 pub mod sigverify_stage;
 pub mod status_deque;
